requirements.txt
tensorflow==2.15.0
deepface==0.0.75
librosa==0.10.1
moviepy==1.0.3
openai-whisper==20231117
pandas==1.5.3
opencv-python==4.8.1.78
numpy==1.23.5
streamlit==1.32.0
scikit-learn==1.3.0
soundfile==0.12.1
ffmpeg-python==0.2.0
protobuf>=3.20,<5.0 